+++ src/main/java/Threaded/Main.java
@@ -347,7 +347,7 @@ public class Main {
             }
 
             // Training loop
-            long trainingStart = System.nanoTime();
+            double trainingStart = System.nanoTime();
             int numThreads = Runtime.getRuntime().availableProcessors();
             CNN cnn = new CNN(numThreads);
             int epochs = 10;
@@ -362,7 +362,7 @@ public class Main {
                 }
                 //System.out.println("Epoch " + epoch + " - Loss: " + totalLoss / trainData.length);
             }
-            long trainingEnd = System.nanoTime();
+            double trainingEnd = System.nanoTime();
             double trainingDuration = (double) (trainingEnd - trainingStart) / 1_000_000;  // Convert to milliseconds
             trainingThreFile.write(String.valueOf(trainingDuration));
 
@@ -379,12 +379,12 @@ public class Main {
             }
 
             // Classify test data
-            long testingStart = System.nanoTime();
+            double testingStart = System.nanoTime();
             for (int i = 0; i < testData.length; i++) {
                 int predictedLabel = cnn.classify(testData[i]);
                 //System.out.println("Sample " + i + " - Predicted Label: " + predictedLabel);
             }
-            long testingEnd = System.nanoTime();
+            double testingEnd = System.nanoTime();
             double testingDuration = (double) (testingEnd - testingStart) / 1_000_000;  // Convert to milliseconds
             testingThreFile.write(String.valueOf(testingDuration));
